requirements.txt
google-generativeai>=0.8.0
scipy>=1.13.0
numpy>=1.26.4
git+https://github.com/suno-ai/bark.git